refactor: moviendo logica y config de Redis a su módulo
