+++ src/main/java/com/l10s/testredis/redis/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.l10s.testredis.config;
+package com.l10s.testredis.redis;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
